pyproject.toml
[tool.poetry]
name = "gradio-demo"
version = "0.1.0"
description = ""
authors = ["Sam Keen <sam.sjk@gmail.com>"]
readme = "README.md"
packages = [{include = "gradio_demo"}]

[tool.poetry.dependencies]
python = "^3.11"
gradio = "^3.39.0"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
trafilatura = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
